refactor: merge 에 따른 디렉토리 구조 변경 - domain 에서 기능별로 분기
